Move update_domain_metrics and run_datadog_user_stats to data_analytics app
